requirements.txt
astroid==1.6.4
autopep8==1.3.5
httplib2==0.11.3
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.4.0
pylint==1.9.1
python-dotenv==0.8.2
six==1.11.0
wrapt==1.10.11
